if文をorElseThrowへ変更・格ファイルをパッケージへ格納
