--- src/main/java/com/u1/user/UserService.java
@@ -1,24 +1,0 @@
-package com.u1.user;
-
-import org.springframework.stereotype.Service;
-
-import java.util.Optional;
-
-@Service
-public class UserService {
-    private final UserMapper userMapper;
-
-    public UserService(UserMapper userMapper) {
-        this.userMapper = userMapper;
-    }
-
-    public User findUser(int id) {
-        Optional<User> user = this.userMapper.findById(id);
-        if (user.isPresent()) {
-            return user.get();
-        } else {
-            throw new UserNotFoundException("user not found");
-        }
-
-    }
-}
