+++ src/main/java/com/u1/user/controller/response/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package com.u1.user;
+package com.u1.user.controller.response;
 
 public class UserNotFoundException extends RuntimeException {
     public UserNotFoundException(String message) {
