+++ src/main/java/com/u1/user/service/UserService.java
@@ -1,0 +1,21 @@
+package com.u1.user.service;
+
+import com.u1.user.controller.response.UserNotFoundException;
+import com.u1.user.entity.User;
+import com.u1.user.mapper.UserMapper;
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserService {
+    private final UserMapper userMapper;
+
+    public UserService(UserMapper userMapper) {
+        this.userMapper = userMapper;
+    }
+
+    public User findUser(int id) {
+        return this.userMapper.findById(id)
+                .orElseThrow(() -> new UserNotFoundException("User not found with id: " + id));
+    }
+}
+
